requirements.txt
stldecompose
numpy
scipy
scikit-learn
matplotlib
pandas
pandas_datareader
investpy
seaborn
xgboost
telebot
telethon
apscheduler
yfinance
plotly
kaleido
ipywidgets
#MetaTrader5
keras
tensorflow-macos